Refactor bank tests
